Refactor indicators into structs
